requirements.txt
langchain
langchain-core

langchain-xai
langchain-openai
dotenv

langchain-huggingface
pygithub
streamlit
GitPython



